% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_functions.R
\name{getSymbolsOfIndex}
\alias{getSymbolsOfIndex}
\title{Get Dataframe of Symbols and Names}
\usage{
getSymbolsOfIndex(index = "atx")
}
\arguments{
\item{index}{(string) index that you want to receive the components of (default is atx)}
}
\description{
This function allows you to get a data frame of components (symbol, company, last_price, change, change_pr) of an index.
The possible indices are ATX, SMI or DAX. Every other text as input leads to an error.
}
\examples{
getSymbolsOfIndex("atx")

}
\keyword{index,}
\keyword{stock,}
\keyword{symbol,}
\keyword{yahoo}
